Consolidate the touch event handling between the two SurfaceView classes
