+++ android/src/org/ppsspp/ppsspp/NativeApp.java
@@ -2,7 +2,18 @@ package org.ppsspp.ppsspp;
 
 // Note that the display* methods are in NativeRenderer.java
 
+import android.annotation.SuppressLint;
+import android.os.Build;
+import android.util.Log;
+import android.view.InputDevice;
+import android.view.InputEvent;
+import android.view.MotionEvent;
+
+import androidx.annotation.RequiresApi;
+
 public class NativeApp {
+	private static final String TAG = "PPSSPPNativeActivity";
+
 	public static final int DEVICE_ID_DEFAULT = 0;
 	public static final int DEVICE_ID_KEYBOARD = 1;
 	public static final int DEVICE_ID_MOUSE = 2;
@@ -83,5 +94,99 @@ public class NativeApp {
 
 	public static void reportError(String errorStr) {
 		NativeApp.sendMessageFromJava("exception", errorStr);
+	}
+
+	@RequiresApi(Build.VERSION_CODES.N)
+	private static void processMouseDelta(final MotionEvent ev) {
+		if ((ev.getSource() & InputDevice.SOURCE_MOUSE) == InputDevice.SOURCE_MOUSE) {
+			float dx = ev.getAxisValue(MotionEvent.AXIS_RELATIVE_X);
+			float dy = ev.getAxisValue(MotionEvent.AXIS_RELATIVE_Y);
+			Log.i(TAG, "Mouse delta: " + dx + " " + dy);
+			NativeApp.mouseDelta(dx, dy);
+		}
+	}
+
+	public static boolean isFromSource(final InputEvent ev, int source) {
+		return (ev.getSource() & source) == source;
+	}
+
+	private static void onMouseEventMotion(final MotionEvent ev) {
+		Log.i(TAG, "motion mouse event");
+		switch (ev.getActionMasked()) {
+			case MotionEvent.ACTION_DOWN: {
+				if (NativeActivity.useModernMouseEvents) {
+					return;
+				}
+				//Log.i(TAG, "Surface Action down. button state: " + ev.getButtonState());
+				NativeApp.mouse(ev.getX(), ev.getY(), 1, 1);
+				break;
+			}
+			case MotionEvent.ACTION_UP: {
+				if (NativeActivity.useModernMouseEvents) {
+					return;
+				}
+				//Log.i(TAG, "Surface Action up. button state: " + ev.getButtonState());
+				NativeApp.mouse(ev.getX(), ev.getY(), 1, 2);
+				break;
+			}
+			case MotionEvent.ACTION_MOVE: {
+				// This still needs handling here, even if new events are used.
+				//Log.i(TAG, "Surface Action move. button state: " + ev.getButtonState());
+				NativeApp.mouse(ev.getX(), ev.getY(), 0, 0);
+				break;
+			}
+			default: {
+				Log.i(TAG, "Unhandled modern mouse action: " + ev.getAction());
+				break;
+			}
+		}
+	}
+
+	// Forwarded from onTouchEvent in the SurfaceView classes.
+	@SuppressLint("ClickableViewAccessibility")
+	public static void processTouchEvent(final MotionEvent ev) {
+		if (isFromSource(ev, InputDevice.SOURCE_MOUSE)) {
+			// This is where workable mouse support arrived.
+			onMouseEventMotion(ev);
+		}
+
+		// Log.i(TAG, "processing touch event");
+		for (int i = 0; i < ev.getPointerCount(); i++) {
+			int pid = ev.getPointerId(i);
+			int code = 0;
+
+			final int action = ev.getActionMasked();
+
+			// These code bits are now the same as the constants in input_state.h.
+			switch (action) {
+				case MotionEvent.ACTION_DOWN:
+				case MotionEvent.ACTION_POINTER_DOWN:
+					// Log.i(TAG, "ACTION_DOWN");
+					if (ev.getActionIndex() == i)
+						code = 2;
+					break;
+				case MotionEvent.ACTION_UP:
+				case MotionEvent.ACTION_POINTER_UP:
+					// Log.i(TAG, "ACTION_UP");
+					if (ev.getActionIndex() == i)
+						code = 4;
+					break;
+				case MotionEvent.ACTION_MOVE: {
+					code = 1;
+					if (Build.VERSION.SDK_INT >= 24) {
+						processMouseDelta(ev);
+					}
+					break;
+				}
+				default:
+					break;
+			}
+
+			if (code != 0) {
+				int tool = ev.getToolType(i);
+				code |= tool << 10; // We use the Android tool type codes
+				NativeApp.touch(ev.getX(i), ev.getY(i), code, pid);
+			}
+		}
 	}
 }
